move openFile and openInFileBrowser functions to proc.cpp utils
